+++ src/hts/private/hts_concat.h
@@ -707,3 +707,13 @@ bcf_fmt_t *bcf_get_fmt(const bcf_hdr_t *hdr, bcf1_t *line, const char *key);
 bcf_info_t *bcf_get_info(const bcf_hdr_t *hdr, bcf1_t *line, const char *key);
 int bcf_get_info_values(const bcf_hdr_t *hdr, bcf1_t *line, const char *tag, void **dst, int *ndst, int type);
 
+
+/**
+*  bcf_hdr_get_hrec() - get header line info
+*  @param type:  one of the BCF_HL_* types: FLT,INFO,FMT,CTG,STR,GEN
+*  @param key:   the header key for generic lines (e.g. "fileformat"), any field
+*                  for structured lines, typically "ID".
+*  @param value: the value which pairs with key. Can be be NULL for BCF_HL_GEN
+*  @param str_class: the class of BCF_HL_STR line (e.g. "ALT" or "SAMPLE"), otherwise NULL
+*/
+bcf_hrec_t *bcf_hdr_get_hrec(const bcf_hdr_t *hdr, int type, const char *key, const char *value, const char *str_class);